+++ src/constants/HeadArticleSearchConstants.js
@@ -1,0 +1,5 @@
+const constants = {
+  placeholderInput: "Search",
+};
+
+export default constants;
