+++ src/styles/_variables.scss
@@ -22,14 +22,13 @@ $widthForContentSecondTabletScreen: 560px;
 
 $transparentBlackBackground: rgba(0, 0, 0, 0.33);
 $headerBackground: rgba(247, 244, 244, 0.4);
+$mainSearchGrey: rgba(196, 196, 196, 0.46);
 $transparentGreykBackground: rgba(116, 113, 113, 0.4);
-
 
 $defaultBorderRadius: 20px;
 $buttonBorderRadius: 10px;
 
 $mobile-breakpoint: 426px;
-
 
 $fontsize-10: 10px;
 $fontsize-12: 12px;
@@ -37,7 +36,3 @@ $fontsize-14: 14px;
 $fontsize-20: 20px;
 $fontsize-24: 24px;
 $bold: 900;
-
-
-
-
